// frontend/src/app/models/gost.ts
export class Gost{
        ime: string="";
        prezime: string="";
        kor_ime: string="";
        pol: string="";
        lozinka: string="";
        bezb_pitanje="";
        bezb_odgovor="";
        telefon: number=0;
        adresa:string="";
        mejl: string="";
        img:string="";
        broj_kartice:number=0;
        status:string="";
}